Move consumers to the folder
